Server/tests/api/requirements.txt
requests==2.24.0
pytest==6.1.1
pydash==4.8.0
python-dateutil==2.8.1
snapshottest==0.6.0
